Shared:Article: rewrite to typescript
