--- packages/management/src/Layout/Article/index.ts
@@ -1,2 +1,0 @@
-export { default as ArticleHeaderContent } from "./Header";
-export { default as ArticleBodyContent } from "./Body";
