+++ packages/shared/components/article/Article.constants.ts
@@ -1,0 +1,3 @@
+export const HEADER_NAME = "Header";
+export const MAIN_BUTTON_NAME = "MainButton";
+export const BODY_NAME = "Body";
