+++ packages/shared/components/article/sub-components/PaymentAlert.tsx
@@ -30,7 +30,7 @@ const ArticlePaymentAlert = ({
       "/payments/portal-payments",
     );
     navigate(paymentPageUrl);
-    toggleArticleOpen();
+    toggleArticleOpen?.();
   };
 
   const title = isFreeTariff ? (
