+++ packages/shared/components/avatar/Avatar.types.ts
@@ -28,5 +28,5 @@ export interface AvatarProps {
   withTooltip?: boolean;
   /** Tooltip content */
   tooltipContent?: string;
-  onClick?: () => void;
+  onClick?: (e: React.MouseEvent) => void;
 }
