+++ packages/shared/components/zendesk/index.ts
@@ -1,4 +1,4 @@
-import { useEffect } from "react";
+import { useEffect, useCallback } from "react";
 
 let waitingChanges: string[][] = [];
 
@@ -7,9 +7,10 @@ const canUseDOM = (): boolean =>
 
 export const ZendeskAPI: Function = (...args: string[]) => {
   if (canUseDOM() && window?.zE) {
+    // @ts-expect-error its ok
     window?.zE?.apply(null, args);
   } else {
-    //console.warn("Zendesk is not initialized yet");
+    // console.warn("Zendesk is not initialized yet");
     waitingChanges.push(args);
   }
 };
@@ -18,11 +19,11 @@ interface Props {
   zendeskKey: string;
   defer?: boolean;
   onLoaded?: () => void;
-  config?: any;
+  config?: {};
 }
 
 const Zendesk = ({ zendeskKey, defer, onLoaded, config }: Props) => {
-  const onScriptLoaded = () => {
+  const onScriptLoaded = useCallback(() => {
     if (waitingChanges.length > 0) {
       waitingChanges.forEach((v) => ZendeskAPI(...v));
       waitingChanges = [];
@@ -31,20 +32,23 @@ const Zendesk = ({ zendeskKey, defer, onLoaded, config }: Props) => {
     if (typeof onLoaded === "function") {
       onLoaded();
     }
-  };
+  }, [onLoaded]);
 
-  const insertScript = (zdKey: string, d?: boolean) => {
-    const script = document.createElement("script");
-    if (d) {
-      script.defer = true;
-    } else {
-      script.async = true;
-    }
-    script.id = "ze-snippet";
-    script.src = `https://static.zdassets.com/ekr/snippet.js?key=${zdKey}`;
-    script.addEventListener("load", onScriptLoaded);
-    document.body.appendChild(script);
-  };
+  const insertScript = useCallback(
+    (zdKey: string, d?: boolean) => {
+      const script = document.createElement("script");
+      if (d) {
+        script.defer = true;
+      } else {
+        script.async = true;
+      }
+      script.id = "ze-snippet";
+      script.src = `https://static.zdassets.com/ekr/snippet.js?key=${zdKey}`;
+      script.addEventListener("load", onScriptLoaded);
+      document.body.appendChild(script);
+    },
+    [onScriptLoaded],
+  );
 
   useEffect(() => {
     if (canUseDOM() && !window?.zE) {
@@ -55,11 +59,13 @@ const Zendesk = ({ zendeskKey, defer, onLoaded, config }: Props) => {
 
     return () => {
       if (canUseDOM() && window.zE) {
+        // @ts-expect-error its ok
         delete window.zE;
+        // @ts-expect-error its ok
         delete window.zESettings;
       }
     };
-  }, [zendeskKey, defer]);
+  }, [zendeskKey, defer, insertScript, config]);
 
   return null;
 };
