+++ packages/shared/skeletons/article/ArticleProfileLoader/ProfileLoader.types.ts
@@ -1,0 +1,8 @@
+export interface ProfileLoaderProps {
+  id?: string;
+  className?: string;
+  style?: React.CSSProperties;
+  showText?: boolean;
+  isVisitor?: boolean;
+  isVirtualKeyboardOpen?: boolean;
+}
