--- packages/shared/skeletons/article/ArticleProfileLoader/index.js
@@ -1,1 +1,0 @@
-export default from "./ArticleProfileLoader";
