+++ packages/shared/skeletons/article/ArticleProfileLoader/index.ts
@@ -1,0 +1,3 @@
+import ArticleProfileLoader from "./ProfileLoader";
+
+export { ArticleProfileLoader };
