+++ packages/shared/skeletons/article/index.ts
@@ -1,3 +1,4 @@
 import { ArticleHeaderLoader } from "./ArticleHeaderLoader";
+import { ArticleProfileLoader } from "./ArticleProfileLoader";
 
-export { ArticleHeaderLoader };
+export { ArticleHeaderLoader, ArticleProfileLoader };
